refactor: move MeetingsByMonthChart to components/, align with project structure
- Move chart to DashboardPage/components/ matching AvgDurationLineChart pattern
- Fix relative import paths in component and test files
- Normalize Stack spacing to {4} to match development branch convention
