+++ src/pages/DashboardPage/SalesmanDashboard/SalesmanDashboard.tsx
@@ -2,7 +2,7 @@ import { EpageDescriptions } from '@data/enums/EpageDescriptions';
 import { EPageTitles } from '@data/enums/EPageTitles';
 import { Stack } from '@mui/material';
 import { AvgDurationLineChart } from '@pages/DashboardPage/components/AvgDurationLineChart/AvgDurationLineChart';
-import { MeetingsByMonthChart } from '@pages/DashboardPage/MeetingsByMonthChart/MeetingsByMonthChart';
+import { MeetingsByMonthChart } from '@pages/DashboardPage/components/MeetingsByMonthChart/MeetingsByMonthChart';
 import { PageContainter } from '@UI/PageContainer/PageContainer';
 import { PageHeader } from '@UI/PageHeader/PageHeader';
 import type { JSX } from 'react';
